requirements.txt
# Web & Server
flask==2.3.3
gunicorn
supervisor
flask-cors==4.0.0

# Core Utilities
requests==2.31.0
pyyaml==6.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# PDF & Web Scraping
PyPDF2==3.0.1
pdfplumber==0.10.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-pptx

# LLM & Vector Tools
torch>=2.1.0
transformers>=4.39.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
langchain==0.1.16
langchain-community==0.0.34
langchain-huggingface>=0.0.3
langchain-text-splitters==0.0.1
chromadb==0.4.6

# APIs
newsapi-python==0.2.6
yfinance==0.2.28